Update tests to handle Sam Phase 1 (#946)
* update tests to handle sam phase 1
