+++ automation/docker/application.conf.ctmpl
@@ -7,6 +7,7 @@ fireCloud {
   baseUrl = "https://firecloud-fiab.dsde-{{$environment}}.broadinstitute.org:22443/"
   orchApiUrl = "https://firecloud-orchestration-fiab.dsde-{{$environment}}.broadinstitute.org:23443/"
   rawlsApiUrl = "https://rawls-fiab.dsde-{{$environment}}.broadinstitute.org:24443/"
+  samApiUrl = "https://sam-fiab.dsde-{{$environment}}.broadinstitute.org:29443/"
   thurloeApiUrl = "https://thurloe-fiab.dsde-{{$environment}}.broadinstitute.org:25443/"
   fireCloudId = "{{$secrets.Data.firecloud_id}}"
 }
