+++ automation/docker/run-tests.sh
@@ -80,6 +80,7 @@ docker run -e DOCKERHOST=$DOCKERHOST \
     --add-host=firecloud-fiab.dsde-${ENV}.broadinstitute.org:${DOCKERHOST} \
     --add-host=firecloud-orchestration-fiab.dsde-${ENV}.broadinstitute.org:${DOCKERHOST} \
     --add-host=rawls-fiab.dsde-${ENV}.broadinstitute.org:${DOCKERHOST} \
+    --add-host=sam-fiab.dsde-${ENV}.broadinstitute.org:${DOCKERHOST} \
     --add-host=thurloe-fiab.dsde-${ENV}.broadinstitute.org:${DOCKERHOST} \
     -P --rm -t -e CHROME_URL="http://hub:4444/" \
     -v $WORKING_DIR/target/application.conf:/app/src/test/resources/application.conf \
