+++ automation/src/test/scala/org/broadinstitute/dsde/firecloud/api/Sam.scala
@@ -1,0 +1,31 @@
+package org.broadinstitute.dsde.firecloud.api
+
+import akka.http.scaladsl.model.StatusCodes
+import com.typesafe.scalalogging.LazyLogging
+import org.broadinstitute.dsde.firecloud.config.{AuthToken, Config}
+
+/**
+  * Sam API service client. This should only be used when Orchestration does
+  * not provide a required endpoint. This should primarily be used for admin
+  * functions.
+  */
+object Sam extends FireCloudClient with LazyLogging {
+
+  private val url = Config.FireCloud.samApiUrl
+
+  object admin {
+
+    def deleteUser(subjectId: String)(implicit token: AuthToken): Unit = {
+      logger.info(s"Deleting user: $subjectId")
+      deleteRequest(url + s"api/admin/user/$subjectId")
+    }
+
+    def doesUserExist(subjectId: String)(implicit token: AuthToken): Option[Boolean] = {
+      getRequest(url + s"api/admin/user/$subjectId").status match {
+        case StatusCodes.OK => Option(true)
+        case StatusCodes.NotFound => Option(false)
+        case _ => None
+      }
+    }
+  }
+}
